Support finding llvm-profdata & llvm-cov with cargo-binutils
Currently, grcov fails with:

  [ERROR] Error while executing llvm tools: We couldn't find llvm-profdata. Try installing the llvm-tools component with `rustup component add llvm-tools-preview`.

This is due to the application using cargo-binutils which looks for
llvm-profdata (and later llvm-cov) in
<rustc-sysroot>/lib/rustlib/<host-triple>/bin. [1]

Address this by adding llvm-profdata and llvm-cov links that allow
applications using cargo-binutils to find the tools without
workarounds. The targets of these links are provided by llvm-16 which
is recommended, but not a dependency. There is precedence for this
already though in rustc in the form of the rust-llvm-dwp link whose
target is also provided by llvm-16.

[1]: https://github.com/rust-embedded/cargo-binutils/blob/master/src/rustc.rs#L15