rust-src: ship original Cargo.lock file
since 1.82.0, rust-analyzer unconditionally runs `cargo metadata` in
sysroot_src (/usr/lib/rustlib/src/library) to discover its contents. since that
directory is owned by root and not writable for regular users, the Cargo.lock
file stored there needs to work "as-is".

because vendored crates are heavily patched for building, but not part of
rust-src, the Cargo.lock file generated as part of the build doesn't match the
shipped rust-src contents. the original one does however, so ship that instead.

Signed-off-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>